+++ external-crates/move/crates/move-compiler/src/expansion/mod.rs
@@ -8,6 +8,7 @@ pub mod ast;
 mod byte_string;
 mod hex_string;
 mod legacy_aliases;
+pub mod name_validation;
 mod path_expander;
 mod primitive_definers;
 pub(crate) mod translate;
